add faction based world map icon, fix lfg button and icons
An error was shown when talking to a flight master, should be fixed now.
Also LFG button is fixed.. but I feel that it is pretty much useless...
World quests that are bosses now displays a skull on blizzard icon type.
Added a new option (Faction) for world quest icon type.